issue 1951 fixes

+++ participant-manager-module/participant-manager-service/src/main/java/com/google/cloud/healthcare/fdamystudies/service/ManageUserServiceImpl.java
@@ -159,7 +159,8 @@ public class ManageUserServiceImpl implements ManageUserService {
     }
 
     if (!user.isSuperAdmin()
-        && (CollectionUtils.isEmpty(user.getApps()) || !hasAtleastOnePermission(user))) {
+        && (user.getManageLocations() == Permission.NO_PERMISSION.value()
+            && (CollectionUtils.isEmpty(user.getApps()) || !hasAtleastOnePermission(user)))) {
       return ErrorCode.PERMISSION_MISSING;
     }
 
@@ -212,25 +213,27 @@ public class ManageUserServiceImpl implements ManageUserService {
         UserMapper.fromUserRequest(user, Long.valueOf(appConfig.getSecurityCodeExpireDate()));
     adminDetails = userAdminRepository.saveAndFlush(adminDetails);
 
-    Map<Boolean, List<UserAppPermissionRequest>> groupBySelectedAppMap =
-        user.getApps()
-            .stream()
-            .collect(Collectors.groupingBy(UserAppPermissionRequest::isSelected));
+    if (!CollectionUtils.isEmpty(user.getApps())) {
+      Map<Boolean, List<UserAppPermissionRequest>> groupBySelectedAppMap =
+          user.getApps()
+              .stream()
+              .collect(Collectors.groupingBy(UserAppPermissionRequest::isSelected));
 
-    // save permissions for selected apps
-    for (UserAppPermissionRequest app :
-        CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.SELECTED))) {
-      saveAppStudySitePermissions(user, adminDetails, app);
-    }
+      // save permissions for selected apps
+      for (UserAppPermissionRequest app :
+          CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.SELECTED))) {
+        saveAppStudySitePermissions(user, adminDetails, app);
+      }
 
-    // save permissions for unselected apps
-    for (UserAppPermissionRequest app :
-        CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.UNSELECTED))) {
-      for (UserStudyPermissionRequest study : CollectionUtils.emptyIfNull(app.getStudies())) {
-        if (study.isSelected()) {
-          saveStudySitePermissions(user, adminDetails, study);
-        } else if (CollectionUtils.isNotEmpty(study.getSites())) {
-          saveSitePermissions(user, adminDetails, study);
+      // save permissions for unselected apps
+      for (UserAppPermissionRequest app :
+          CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.UNSELECTED))) {
+        for (UserStudyPermissionRequest study : CollectionUtils.emptyIfNull(app.getStudies())) {
+          if (study.isSelected()) {
+            saveStudySitePermissions(user, adminDetails, study);
+          } else if (CollectionUtils.isNotEmpty(study.getSites())) {
+            saveSitePermissions(user, adminDetails, study);
+          }
         }
       }
     }
@@ -365,7 +368,7 @@ public class ManageUserServiceImpl implements ManageUserService {
 
     AdminUserResponse userResponse =
         user.isSuperAdmin()
-            ? updateSuperAdminDetails(user, superAdminUserId, auditRequest)
+            ? updateSuperAdminDetails(user, auditRequest)
             : updateAdminDetails(user, superAdminUserId, auditRequest);
     String accessLevel = user.isSuperAdmin() ? CommonConstants.SUPER_ADMIN : CommonConstants.ADMIN;
     if (MessageCode.UPDATE_USER_SUCCESS.getMessage().equals(userResponse.getMessage())) {
@@ -395,7 +398,9 @@ public class ManageUserServiceImpl implements ManageUserService {
       return ErrorCode.USER_NOT_FOUND;
     }
 
-    if (!user.isSuperAdmin() && !hasAtleastOnePermission(user)) {
+    if (!user.isSuperAdmin()
+        && (user.getManageLocations() == Permission.NO_PERMISSION.value()
+            && (CollectionUtils.isEmpty(user.getApps()) || !hasAtleastOnePermission(user)))) {
       return ErrorCode.PERMISSION_MISSING;
     }
     logger.exit("Successfully validated user request");
@@ -403,7 +408,7 @@ public class ManageUserServiceImpl implements ManageUserService {
   }
 
   private AdminUserResponse updateSuperAdminDetails(
-      UserRequest user, String superAdminUserId, AuditLogEventRequest auditRequest) {
+      UserRequest user, AuditLogEventRequest auditRequest) {
     logger.entry("updateSuperAdminDetails()");
     Optional<UserRegAdminEntity> optAdminDetails = userAdminRepository.findById(user.getId());
 
@@ -462,29 +467,30 @@ public class ManageUserServiceImpl implements ManageUserService {
     adminDetails = UserMapper.fromUpdateUserRequest(user, adminDetails);
     userAdminRepository.saveAndFlush(adminDetails);
 
-    deleteAllPermissions(user.getId());
+    if (CollectionUtils.isNotEmpty(user.getApps())) {
+      deleteAllPermissions(user.getId());
+      user.setSuperAdminUserId(superAdminUserId);
 
-    user.setSuperAdminUserId(superAdminUserId);
+      Map<Boolean, List<UserAppPermissionRequest>> groupBySelectedAppMap =
+          user.getApps()
+              .stream()
+              .collect(Collectors.groupingBy(UserAppPermissionRequest::isSelected));
 
-    Map<Boolean, List<UserAppPermissionRequest>> groupBySelectedAppMap =
-        user.getApps()
-            .stream()
-            .collect(Collectors.groupingBy(UserAppPermissionRequest::isSelected));
+      // save permissions for selected apps
+      for (UserAppPermissionRequest app :
+          CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.SELECTED))) {
+        saveAppStudySitePermissions(user, adminDetails, app);
+      }
 
-    // save permissions for selected apps
-    for (UserAppPermissionRequest app :
-        CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.SELECTED))) {
-      saveAppStudySitePermissions(user, adminDetails, app);
-    }
-
-    // save permissions for unselected apps
-    for (UserAppPermissionRequest app :
-        CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.UNSELECTED))) {
-      for (UserStudyPermissionRequest study : CollectionUtils.emptyIfNull(app.getStudies())) {
-        if (study.isSelected()) {
-          saveStudySitePermissions(user, adminDetails, study);
-        } else if (CollectionUtils.isNotEmpty(study.getSites())) {
-          saveSitePermissions(user, adminDetails, study);
+      // save permissions for unselected apps
+      for (UserAppPermissionRequest app :
+          CollectionUtils.emptyIfNull(groupBySelectedAppMap.get(CommonConstants.UNSELECTED))) {
+        for (UserStudyPermissionRequest study : CollectionUtils.emptyIfNull(app.getStudies())) {
+          if (study.isSelected()) {
+            saveStudySitePermissions(user, adminDetails, study);
+          } else if (CollectionUtils.isNotEmpty(study.getSites())) {
+            saveSitePermissions(user, adminDetails, study);
+          }
         }
       }
     }
